requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langchain-community>=0.0.20
mcp>=1.0.0
google-generativeai>=0.8.0
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
chromadb>=0.5.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
